requirements.txt
flask
psycopg2-binary
Flask-SQLAlchemy
PyJWT==1.7.1
